+++ internal/http/www/static/fs.go
@@ -5,5 +5,5 @@ package static
 
 import "embed"
 
-//go:embed *.css *.png */*.ttf *.webmanifest
+//go:embed *.css *.js *.png *.ttf *.woff2 *.webmanifest
 var FS embed.FS